Lidar system simulation/lidar/processor.py
"""
Point cloud processing module for LiDAR simulation.
Handles clustering, outlier removal, and denoising.
"""

import numpy as np
from sklearn.cluster import DBSCAN
from typing import Dict, Any, Tuple

class PointCloudProcessor:
    """Processor for LiDAR point cloud data."""
    
    def __init__(self, **kwargs):
        """
        Initialize the point cloud processor.
        
        Args:
            **kwargs: Processing parameters
        """
        self.clustering_params = kwargs.get('clustering', {})
        self.outlier_params = kwargs.get('outlier_removal', {})
        self.denoising_params = kwargs.get('denoising', {})
    
    def process(self, points: np.ndarray, intensities: np.ndarray) -> np.ndarray:
        """
        Process point cloud data through the pipeline.
        
        Args:
            points: Nx3 array of 3D points
            intensities: N array of intensity values
            
        Returns:
            Processed point cloud
        """
        # Remove outliers if enabled
        if self.outlier_params.get('enabled', False):
            points = self._remove_outliers(points)
        
        # Apply denoising if enabled
        if self.denoising_params.get('enabled', False):
            points = self._denoise(points)
        
        # Apply clustering if enabled
        if self.clustering_params:
            points = self._cluster_points(points)
        
        return points
    
    def _remove_outliers(self, points: np.ndarray) -> np.ndarray:
        """
        Remove statistical outliers from the point cloud.
        
        Args:
            points: Nx3 array of 3D points
            
        Returns:
            Filtered point cloud
        """
        # Calculate mean and standard deviation
        mean = np.mean(points, axis=0)
        std = np.std(points, axis=0)
        
        # Calculate distances from mean
        distances = np.linalg.norm(points - mean, axis=1)
        
        # Filter points within standard deviation threshold
        threshold = self.outlier_params.get('std_dev', 1.5)
        mask = distances < threshold * np.mean(distances)
        
        return points[mask]
    
    def _denoise(self, points: np.ndarray) -> np.ndarray:
        """
        Apply denoising to the point cloud.
        
        Args:
            points: Nx3 array of 3D points
            
        Returns:
            Denoised point cloud
        """
        method = self.denoising_params.get('method', 'statistical')
        k_neighbors = self.denoising_params.get('k_neighbors', 20)
        
        if method == 'statistical':
            return self._statistical_denoising(points, k_neighbors)
        else:
            return self._bilateral_denoising(points)
    
    def _statistical_denoising(self, points: np.ndarray, k: int) -> np.ndarray:
        """
        Apply statistical denoising.
        
        Args:
            points: Nx3 array of 3D points
            k: Number of neighbors to consider
            
        Returns:
            Denoised point cloud
        """
        # Create KD-tree for efficient neighbor search
        from scipy.spatial import cKDTree
        tree = cKDTree(points)
        
        # Find k-nearest neighbors for each point
        distances, indices = tree.query(points, k=k)
        
        # Calculate mean and standard deviation for each neighborhood
        means = np.mean(points[indices], axis=1)
        stds = np.std(distances, axis=1)
        
        # Filter points based on distance from mean
        mask = np.linalg.norm(points - means, axis=1) < 1.5 * stds[:, np.newaxis]
        
        return points[mask]
    
    def _bilateral_denoising(self, points: np.ndarray) -> np.ndarray:
        """
        Apply bilateral filtering for denoising.
        
        Args:
            points: Nx3 array of 3D points
            
        Returns:
            Denoised point cloud
        """
        # Simple implementation - can be enhanced with proper bilateral filtering
        from scipy.ndimage import gaussian_filter
        return gaussian_filter(points, sigma=1.0)
    
    def _cluster_points(self, points: np.ndarray) -> np.ndarray:
        """
        Apply clustering to the point cloud.
        
        Args:
            points: Nx3 array of 3D points
            
        Returns:
            Clustered point cloud
        """
        # Apply DBSCAN clustering
        clustering = DBSCAN(
            eps=self.clustering_params.get('eps', 0.5),
            min_samples=self.clustering_params.get('min_samples', 10)
        ).fit(points)
        
        # Get labels
        labels = clustering.labels_
        
        # Filter out noise points (label = -1)
        mask = labels != -1
        return points[mask]
    
    def get_cluster_labels(self, points: np.ndarray) -> np.ndarray:
        """
        Get cluster labels for the point cloud.
        
        Args:
            points: Nx3 array of 3D points
            
        Returns:
            Array of cluster labels
        """
        clustering = DBSCAN(
            eps=self.clustering_params.get('eps', 0.5),
            min_samples=self.clustering_params.get('min_samples', 10)
        ).fit(points)
        
        return clustering.labels_ 