Lidar system simulation/requirements.txt
trimesh>=3.9.0
numpy>=1.21.0
open3d>=0.13.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
rtree>=1.4.0
pyyaml>=6.0.0
pydantic>=2.0.0
tqdm>=4.65.0
pyembree>=0.1.8  # Optional: for faster ray tracing 